[TASK] Deprecate the `ReadOnly` trait
The new Trait now is named `ReadOnlyRepository`.
(The old name breaks compatibility with PHP 8.1.)

Fixes #621
